Remove sink/api
