--- sinks/types.go
@@ -1,24 +1,0 @@
-// Copyright 2014 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// This package implements the various storage backends supported by heapster.
-package sinks
-
-// TODO: Once in-memory storage is implemented, make the manager extract data from the in-memory store periodically.
-type Sink interface {
-	// Sync starts exporting stats periodically to backends.
-	// Returns a channel that can be used to stop syncing.
-	Sync() chan<- struct{}
-	DebugInfo() string
-}
